+++ javascript/modules/client/src/internal/interfaces/core.ts
@@ -33,9 +33,9 @@ export interface IClientIpfsCore {
   fetchBytes: (cid: string) => Promise<Uint8Array | undefined>;
 }
 export interface IClientGraphQLCore {
-  getClient: () => GraphQLClient
-  shiftClient: () => void
-  isUp: () => Promise<boolean>
+  getClient: () => GraphQLClient;
+  shiftClient: () => void;
+  isUp: () => Promise<boolean>;
   ensureOnline: () => Promise<void>;
 }