+++ javascript/modules/client/src/internal/temp-mock.ts
@@ -6,4 +6,3 @@ export const delay = (delayInms: number) => {
   }
   return new Promise((resolve) => setTimeout(resolve, delay));
 };
-