+++ javascript/modules/client/test/integration/client-addressList.test.ts
@@ -27,6 +27,7 @@ import {
   IVoteProposalParams,
   ProposalCreationSteps,
   ProposalSortBy,
+  ProposalStatus,
   VoteProposalStep,
   VoteValues,
 } from "../../src/internal/interfaces/plugins";
@@ -40,7 +41,9 @@ const IPFS_API_KEY = process.env.IPFS_API_KEY ||
   Buffer.from(
     "YjQ3N1JoRUNmOHM4c2RNN1hya0xCczJ3SGM0a0NNd3BiY0ZDNTVLdCAg==",
     "base64",
-  ).toString().trim();
+  )
+    .toString()
+    .trim();
 
 const web3endpoints = {
   working: [
@@ -68,7 +71,7 @@ const grapqhlEndpoints = {
   working: [
     {
       url:
-        "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-rinkeby",
+        "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-goerli",
     },
   ],
   failing: [{ url: "https://bad-url-gateway.io/" }],
@@ -76,6 +79,11 @@ const grapqhlEndpoints = {
 
 const TEST_WALLET =
   "0xdf57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656e";
+const TEST_PROPOSAL_ID = "0x0bcaeb2da9d8ce246ac0864f081b04d4380b6b9c_0x0";
+const TEST_DAO_ADDDRESS = "0xa40fd495c454d19c87c4d615100307f0283c017c";
+const TEST_PLUGIN_ADDRESS = "0x0bcaeb2da9d8ce246ac0864f081b04d4380b6b9c";
+const TEST_INVALID_ADDRESS = "0x1nv4l1d_4ddr355";
+const TEST_NON_EXISTING_ADDRESS = "0x1234567890123456789012345678901234567890";
 
 const contextParams: ContextParams = {
   network: "mainnet",
@@ -111,10 +119,12 @@ const contextParamsLocalChain: ContextParams = {
       url: "http:localhost:5003",
     },
   ],
-  graphqlNodes: [{
-    url:
-      "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-rinkeby",
-  }],
+  graphqlNodes: [
+    {
+      url:
+        "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-goerli",
+    },
+  ],
 };
 
 describe("Client", () => {
@@ -131,7 +141,7 @@ describe("Client", () => {
   describe("Client instances", () => {
     it("Should create a working client", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       expect(client).toBeInstanceOf(ClientAddressList);
@@ -155,7 +165,7 @@ describe("Client", () => {
 
     it("Should create a failing client", async () => {
       const ctx = new Context(contextParamsFailing);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       expect(client).toBeInstanceOf(ClientAddressList);
@@ -178,7 +188,7 @@ describe("Client", () => {
   describe("Proposal Creation", () => {
     it("Should estimate the gas fees for creating a new proposal", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const proposalParams: ICreateProposalParams = {
@@ -187,10 +197,12 @@ describe("Client", () => {
           title: "Best Proposal",
           summary: "this is the sumnary",
           description: "This is a very long description",
-          resources: [{
-            name: "Website",
-            url: "https://the.website",
-          }],
+          resources: [
+            {
+              name: "Website",
+              url: "https://the.website",
+            },
+          ],
           media: {
             header: "https://no.media/media.jpeg",
             logo: "https://no.media/media.jpeg",
@@ -213,7 +225,7 @@ describe("Client", () => {
     });
     it("Should create a new proposal locally", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const addressListClient = new ClientAddressList(ctxPlugin);
       const client = new Client(ctx);
 
@@ -233,10 +245,12 @@ describe("Client", () => {
           title: "Best Proposal",
           summary: "this is the sumnary",
           description: "This is a very long description",
-          resources: [{
-            name: "Website",
-            url: "https://the.website",
-          }],
+          resources: [
+            {
+              name: "Website",
+              url: "https://the.website",
+            },
+          ],
           media: {
             header: "https://no.media/media.jpeg",
             logo: "https://no.media/media.jpeg",
@@ -250,7 +264,9 @@ describe("Client", () => {
       };
 
       for await (
-        const step of addressListClient.methods.createProposal(proposalParams)
+        const step of addressListClient.methods.createProposal(
+          proposalParams,
+        )
       ) {
         switch (step.key) {
           case ProposalCreationSteps.CREATING:
@@ -274,7 +290,7 @@ describe("Client", () => {
   describe("Vote on a proposal", () => {
     it("Should estimate the gas fees for casting a vote", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const voteParams: IVoteProposalParams = {
@@ -294,7 +310,7 @@ describe("Client", () => {
 
     it("Should vote on a proposal locally", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const voteParams: IVoteProposalParams = {
@@ -325,7 +341,7 @@ describe("Client", () => {
   describe("Execute proposal", () => {
     it("Should estimate the gas fees for executing a proposal", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const executeParams: IExecuteProposalParams = {
@@ -343,7 +359,7 @@ describe("Client", () => {
 
     it("Should execute a local proposal", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const executeParams: IExecuteProposalParams = {
@@ -371,7 +387,7 @@ describe("Client", () => {
   describe("Can vote", () => {
     it("Should check if an user can vote in an Address List proposal", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const params: ICanVoteParams = {
@@ -401,7 +417,9 @@ describe("Client", () => {
       };
 
       const installPluginItemItem = ClientAddressList.encoding
-        .getPluginInstallItem(withdrawParams);
+        .getPluginInstallItem(
+          withdrawParams,
+        );
 
       expect(typeof installPluginItemItem).toBe("object");
       // what does this should be
@@ -410,7 +428,7 @@ describe("Client", () => {
 
     it("Should create an AddressList client and fail to generate a plugin config action with an invalid address", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const pluginConfigParams: IPluginSettings = {
@@ -430,7 +448,7 @@ describe("Client", () => {
 
     it("Should create an AddressList client and generate a plugin config action action", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const pluginConfigParams: IPluginSettings = {
@@ -454,7 +472,7 @@ describe("Client", () => {
 
     it("Should encode a add members action with an invalid plugin address and fail", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -464,16 +482,12 @@ describe("Client", () => {
       ];
 
       const pluginAddress = "0xinvalid_address";
-      expect(() =>
-        client.encoding.addMembersAction(
-          pluginAddress,
-          members,
-        )
-      ).toThrow(new InvalidAddressError());
+      expect(() => client.encoding.addMembersAction(pluginAddress, members))
+        .toThrow(new InvalidAddressError());
     });
     it("Should encode a add members action with an invalid member address and fail", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -483,16 +497,12 @@ describe("Client", () => {
       ];
 
       const pluginAddress = "0x1234567890123456789012345678901234567890";
-      expect(() =>
-        client.encoding.addMembersAction(
-          pluginAddress,
-          members,
-        )
-      ).toThrow(new InvalidAddressError());
+      expect(() => client.encoding.addMembersAction(pluginAddress, members))
+        .toThrow(new InvalidAddressError());
     });
     it("Should encode a add members action", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -513,7 +523,7 @@ describe("Client", () => {
     });
     it("Should encode a remove members action with an invalid plugin address and fail", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -523,16 +533,12 @@ describe("Client", () => {
       ];
 
       const pluginAddress = "0xinvalid_address";
-      expect(() =>
-        client.encoding.removeMembersAction(
-          pluginAddress,
-          members,
-        )
-      ).toThrow(new InvalidAddressError());
+      expect(() => client.encoding.removeMembersAction(pluginAddress, members))
+        .toThrow(new InvalidAddressError());
     });
     it("Should encode a remove members action with an invalid member address and fail", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -542,16 +548,12 @@ describe("Client", () => {
       ];
 
       const pluginAddress = "0x1234567890123456789012345678901234567890";
-      expect(() =>
-        client.encoding.removeMembersAction(
-          pluginAddress,
-          members,
-        )
-      ).toThrow(new InvalidAddressError());
+      expect(() => client.encoding.removeMembersAction(pluginAddress, members))
+        .toThrow(new InvalidAddressError());
     });
     it("Should encode a remove members action", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -578,7 +580,7 @@ describe("Client", () => {
   describe("Action decoders", () => {
     it("Should decode the plugin settings from an update plugin settings action", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const params: IPluginSettings = {
         minDuration: 7200,
@@ -589,9 +591,14 @@ describe("Client", () => {
       const pluginAddress = "0x1234567890123456789012345678901234567890";
 
       const updatePluginSettingsAction = client.encoding
-        .updatePluginSettingsAction(pluginAddress, params);
+        .updatePluginSettingsAction(
+          pluginAddress,
+          params,
+        );
       const decodedParams: IPluginSettings = client.decoding
-        .updatePluginSettingsAction(updatePluginSettingsAction.data);
+        .updatePluginSettingsAction(
+          updatePluginSettingsAction.data,
+        );
 
       expect(decodedParams.minDuration).toBe(params.minDuration);
       expect(decodedParams.minSupport).toBe(params.minSupport);
@@ -600,21 +607,17 @@ describe("Client", () => {
 
     it("Should try to decode a invalid action and with the update plugin settings decoder return an error", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const data = new Uint8Array([11, 22, 22, 33, 33, 33]);
 
-      expect(() =>
-        client.decoding
-          .updatePluginSettingsAction(data)
-      )
-        .toThrow(
-          `no matching function (argument="sighash", value="0x0b161621", code=INVALID_ARGUMENT, version=abi/5.6.0)`,
-        );
+      expect(() => client.decoding.updatePluginSettingsAction(data)).toThrow(
+        `no matching function (argument="sighash", value="0x0b161621", code=INVALID_ARGUMENT, version=abi/5.6.0)`,
+      );
     });
     it("Should decode a add members action", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -634,7 +637,7 @@ describe("Client", () => {
     });
     it("Should decode a remove members action", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const members: string[] = [
@@ -658,7 +661,7 @@ describe("Client", () => {
 
     it("Should get the function for a given action data", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const params: IPluginSettings = {
         minDuration: 7200,
@@ -669,7 +672,10 @@ describe("Client", () => {
       const pluginAddress = "0x1234567890123456789012345678901234567890";
 
       const updatePluginSettingsAction = client.encoding
-        .updatePluginSettingsAction(pluginAddress, params);
+        .updatePluginSettingsAction(
+          pluginAddress,
+          params,
+        );
       const iface = client.decoding.findInterface(
         updatePluginSettingsAction.data,
       );
@@ -680,7 +686,7 @@ describe("Client", () => {
 
     it("Should try to get the function of an invalid data and return null", async () => {
       const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const data = new Uint8Array([11, 22, 22, 33, 33, 33]);
       const iface = client.decoding.findInterface(data);
@@ -691,11 +697,11 @@ describe("Client", () => {
   describe("Data retrieval", () => {
     it("Should get the list of members that can vote in a proposal", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
       const wallets = await client.methods.getMembers(
-        "0x1234567890123456789012345678901234567890",
+        TEST_DAO_ADDDRESS,
       );
 
       expect(Array.isArray(wallets)).toBe(true);
@@ -705,10 +711,10 @@ describe("Client", () => {
     });
     it("Should fetch the given proposal", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
-      const proposalId = "0x8ba1e148c44fbc936dc9a1e06648693a55a16fa4_0x1";
+      const proposalId = TEST_PROPOSAL_ID;
       const proposal = await client.methods.getProposal(proposalId);
 
       expect(typeof proposal).toBe("object");
@@ -782,23 +788,25 @@ describe("Client", () => {
     });
     it("Should fetch the given proposal and fail because the proposal does not exist", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
-      const proposalId = "0x1234567890123456789012345678901234567890_0x0";
+      const proposalId = TEST_NON_EXISTING_ADDRESS + "_0x0";
       const proposal = await client.methods.getProposal(proposalId);
 
       expect(proposal === null).toBe(true);
     });
     it("Should get a list of proposals filtered by the given criteria", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const limit = 5;
+      const status = ProposalStatus.EXECUTED;
       const params: IProposalQueryParams = {
         limit,
         sortBy: ProposalSortBy.CREATED_AT,
         direction: SortDirection.ASC,
+        status,
       };
       const proposals = await client.methods.getProposals(params);
 
@@ -817,6 +825,7 @@ describe("Client", () => {
         expect(typeof proposal.metadata.summary).toBe("string");
         expect(proposal.startDate instanceof Date).toBe(true);
         expect(proposal.endDate instanceof Date).toBe(true);
+        expect(proposal.status).toBe(status);
         // result
         expect(typeof proposal.result.yes).toBe("number");
         expect(typeof proposal.result.no).toBe("number");
@@ -825,10 +834,10 @@ describe("Client", () => {
     });
     it("Should get a list of proposals from a specific dao", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const limit = 5;
-      const address = "0xe800043252a9d7ccff5714998293b742caed2361";
+      const address = TEST_DAO_ADDDRESS;
       const params: IProposalQueryParams = {
         limit,
         sortBy: ProposalSortBy.CREATED_AT,
@@ -842,10 +851,10 @@ describe("Client", () => {
     });
     it("Should get a list of proposals from a dao that has no proposals", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const limit = 5;
-      const address = "0x1234567890123456789012345678901234567890";
+      const address = TEST_NON_EXISTING_ADDRESS;
       const params: IProposalQueryParams = {
         limit,
         sortBy: ProposalSortBy.CREATED_AT,
@@ -859,10 +868,10 @@ describe("Client", () => {
     });
     it("Should get a list of proposals from an invalid address", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
       const limit = 5;
-      const address = "0xn0tv4l1d";
+      const address = TEST_INVALID_ADDRESS;
       const params: IProposalQueryParams = {
         limit,
         sortBy: ProposalSortBy.CREATED_AT,
@@ -875,11 +884,10 @@ describe("Client", () => {
     });
     it("Should get the settings of a plugin given a plugin instance address", async () => {
       const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx)
+      const ctxPlugin = ContextPlugin.fromContext(ctx);
       const client = new ClientAddressList(ctxPlugin);
 
-      const pluginAddress: string =
-        "0x8ba1e148c44fbc936dc9a1e06648693a55a16fa4";
+      const pluginAddress: string = TEST_PLUGIN_ADDRESS;
       const settings = await client.methods.getSettings(pluginAddress);
 
       expect(settings === null).toBe(false);